requirements.txt
websockets
pyserial
pyserial-asyncio
customtkinter
pystray
pillow
pyinstaller
pyserial
pyserial-asyncio
pyQt5
pyQt5-sip
ipdb    